Remove domain file Move enum to file storage file VDX-1934
