--- plato/domain.py
@@ -1,6 +1,0 @@
-from enum import Enum
-
-
-class StorageType(str, Enum):
-    S3 = 's3'
-    DISK = 'disk'
